+++ lib/Service/FileDuplicateService.php
@@ -73,19 +73,19 @@ class FileDuplicateService
         bool $enrich = false,
         ?array $orderBy = [['hash'], ['type']]
     ): array {
-        $limit = $pageSize; // Set the number of records per page
-        $offset = ($page - 1) * $pageSize; // Calculate the offset
+        $result = [];
+        $isLastFetched = false;
+        $entities = [];
 
-        $result = array();
-        $entities = null;
-        do {
-            $entities = $this->mapper->findAll($user, $limit, $offset, $orderBy);
+        while (empty($entities) && !$isLastFetched) {
+            $offset = ($page - 1) * $pageSize; // Calculate the offset based on the current page
+            $entities = $this->mapper->findAll($user, $pageSize, $offset, $orderBy);
+
             foreach ($entities as $entity) {
                 $entity = $this->stripFilesWithoutAccessRights($entity, $user);
                 if ($enrich) {
                     $entity = $this->enrich($entity);
                 }
-                $offset = $entity->id;
                 if (count($entity->getFiles()) > 1) {
                     if ($type === 'acknowledged' && $entity->isAcknowledged()) {
                         $result[] = $entity;
@@ -94,14 +94,20 @@ class FileDuplicateService
                     } else if ($type === 'all') {
                         $result[] = $entity;
                     }
-                    if (count($result) === $limit) {
-                        break;
-                    }
                 }
             }
-            unset($entity);
-        } while (count($result) < $limit && count($entities) === $limit);
-        return array("entities" => $result, "pageKey" => $offset, "isLastFetched" => count($entities) !== $limit);
+
+            $isLastFetched = count($entities) < $pageSize; // Determine if this is the last page
+            if (empty($entities) && !$isLastFetched) {
+                $page++; // Move to the next page if no entities found and not the last page
+            }
+        }
+
+        return [
+            "entities" => $result,
+            "pageKey" => $offset,
+            "isLastFetched" => $isLastFetched
+        ];
     }
 
     private function stripFilesWithoutAccessRights(